Prototyping/pyproject.toml
[project]
name = "prototyping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "duckduckgo-search>=8.1.1",
    "googlesearch-python>=1.3.0",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.1",
    "pydantic>=2.12.5",
    "structlog>=25.5.0",
    "pydantic-settings>=2.12.0",
    "snowflake-connector-python>=3.10.0",
    "pandas>=2.0.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "python-dotenv>=1.0.1",
    "python-docx>=1.2.0",
    "litellm>=1.31.0",
    "chromadb>=0.4.24",
    "sentence-transformers>=2.5.1",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.4.0",
]
